Began writing tests for the scan matcher.
